Correcciones visuales

+++ app/Http/Controllers/EjemplarController.php
@@ -8,30 +8,37 @@ use App\field;
 use App\Http\Controllers\pagesController;
 use App\Media;
 use App\relation;
+use Auth;
 use DB;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 
 class EjemplarController extends Controller
 {
-    // public $notificacion = "";
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @param \Illuminate\Http\Response
      */
 
     public function index(Request $request, $notificacion = false, $admin = false)
     {
         $accionGet = new data_field();
         $listEje = $accionGet->getDataField();
-
         if ($request->ajax()) {
             $filterSex = $request->input('sex');
             $filterName = $request->input('name');
             $filterPagination = $request->input('page');
             $filterColor = $request->input('filterColor');
             $filterRaza = $request->input('raza');
+            $filterRaza = str_replace("%20", " ", $filterRaza);
+            if (strpos($filterRaza, "Bulldog I") !== false) {
+                $filterRaza = "Bulldog Inglés";
+            }
+            if (strpos($filterRaza, "Bulldog F") !== false) {
+                $filterRaza = "Bulldog Francés";
+            }
+
             $perPage = $request->input('perPage');
             $slug = $request->input('slug');
             $admin = $request->input('role');
@@ -60,40 +67,63 @@ class EjemplarController extends Controller
                 $Ejefiltrados = $accionGet->filtroArray($Ejefiltrados, 'raza', $filterRaza);
             }
 
+
             $ejemplares = $accionGet->paginador($Ejefiltrados, $nroPage, $perPage);
             return response()->json(view('public.sublista', compact('ejemplares', 'admin'))->render());
         } else {
+            $user = Auth::user();
+            if ($user) {
+                $razas = DB::table('razas')
+                    ->select('raza')
+                    ->get();
 
-            $razas = DB::table('razas')
-                ->select('raza')
-                ->get();
+                    $media=new Media;
 
-            $ejemplares = $accionGet->paginador($listEje, null, 10);
+                    
 
-            if (strpos(url()->current(), "Ejemplar")) {
-                $admin = true;
+                $ejemplares = $accionGet->paginador($listEje, null, 10);
+
+                if (strpos(url()->current(), "Ejemplar")) {
+                    $admin = true;
+                }
+
+                return view('admin.ejemplars.ejemplares', compact('ejemplares', 'razas', 'notificacion', 'admin'))->render();
+            } else {
+                return redirect('/g/iniciar-sesion');
             }
-
-            return view('admin.ejemplars.ejemplares', compact('ejemplares', 'razas', 'notificacion', 'admin'))->render();
         }
+    }
+
+    /**
+     *Funciòn que devuelve el dashboard 
+     */
+    public function dashboard()
+    {
+        $user = Auth::user();
+            $ejemplares = DB::table('ejemplars')->count();
+            return view("admin.dashboard", compact("ejemplares"));
     }
 
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @param \Illuminate\Http\Response
      */
-    public function create()
+    public function create(Request $request)
     {
+        $user = Auth::user();
+        if ($user) {
 
-        foreach (DB::table('razas')->select('raza')->cursor() as $allRazas) {
-            $razas[$allRazas->raza] = $allRazas->raza;
+            foreach (DB::table('razas')->select('raza')->cursor() as $allRazas) {
+                $razas[$allRazas->raza] = $allRazas->raza;
+            }
+
+            $columns = field::all();
+
+            return view('admin.ejemplars.ejemplar', compact('razas', 'columns'));
+        } else {
+            return redirect('/g/iniciar-sesion');
         }
-
-        $columns = field::all();
-
-        // return $columns;
-        return view('admin.ejemplars.ejemplar', compact('razas', 'columns'));
 
     }
 
@@ -105,39 +135,53 @@ class EjemplarController extends Controller
     public function store(Request $request)
     {
 
-        $datos = [];
-        $media = new Media;
-        $relation = new relation();
-        $ejemplar = new Ejemplar();
+        $user = Auth::user();
+        if ($user) {
+            $datos = [];
+            $media = new Media;
+            $relation = new relation();
+            $ejemplar = new Ejemplar();
 
-        $ejemplar->slug = Str::slug($request['name'], '-') . '-' . rand(1, 99999);
-        $ejemplar->save();
+            $ejemplar->slug = Str::slug($request['name'], '-') . '-' . rand(1, 99999);
+            $ejemplar->save();
+            $id = Ejemplar::latest()
+            ->take(1)
+            ->select("id")
+            ->get();
+            $id=$id[0]->id;
 
-        $id = Ejemplar::all()->last()->id;
+            
+            $media->saveMedia($request, $id);
+            $relation->saveRelations($request, $id);
 
-        $media->saveMedia($request, $id);
-        $relation->saveRelations($request, $id);
+            // Ciclo for para almacenar los datos
+            foreach ($request->request as $key => $value) {
+                $idColumn = DB::table('fields')
+                    ->where('name', $key)
+                    ->value('id');
 
-        // Ciclo for para almacenar los datos
-        foreach ($request->request as $key => $value) {
-            $idColumn = DB::table('fields')
-                ->where('name', $key)
-                ->value('id');
-
-            if (!is_null($idColumn)) {
-                $item['field_id'] = $idColumn;
-                $item['data'] = $value;
-                $item['ejemplar_id'] = $id;
-                array_push($datos, $item);
+                if (!is_null($idColumn)) {
+                    $item['field_id'] = $idColumn;
+                    $item['data'] = $value;
+                    $item['ejemplar_id'] = $id;
+                    array_push($datos, $item);
+                }
             }
+
+            DB::table('data_fields')->insert($datos);
+
+            return redirect('/Ejemplar')->with('status', 'Ejemplar añadido!');
+        } else {
+            return redirect('/g/iniciar-sesion');
         }
-
-        DB::table('data_fields')->insert($datos);
-
-        return redirect('/Ejemplar')->with('status', 'Ejemplar añadido!');
 
     }
 
+    /**
+     * Función que se encarga de traer al lado del cliente todo el arból genealogico de 2 ejemplares
+
+     * @param mixed $params - Los slugs de los ejemplares extraidos desde un get
+     */
     public function simulator($params)
     {
         $ejemplar = new Ejemplar();
@@ -183,7 +227,7 @@ class EjemplarController extends Controller
                 "Hijos" => $hijos],
         ];
 
-        $page = $page->show($req);
+        $page = $page->show();
         $ordenCard = $page["orden"];
 
         foreach ($ordenCard as $key => $value) {
@@ -192,7 +236,14 @@ class EjemplarController extends Controller
             $orden[$value->publicName] = $valor;
 
         }
-        return view('public.ejemplar', compact('details', 'abuelos', 'page', 'orden'));
+        $orden2 = $page["ordenSeeder"];
+        foreach ($orden2 as $key => $value) {
+            $name = $value->columnName;
+            $valor = $ejemplar[$name];
+            $ordenSeeder[$value->publicName] = $valor;
+
+        }
+        return view('public.ejemplar', compact('details', 'abuelos', 'page', 'orden', 'ordenSeeder'));
 
     }
 
@@ -204,21 +255,25 @@ class EjemplarController extends Controller
      */
     public function edit($slug)
     {
+        $user = Auth::user();
+        if ($user) {
+            $datos = new Ejemplar();
+            $id = $datos->getIdEjemplar($slug);
+            $padres = $datos->getParents($id);
 
-        $datos = new Ejemplar();
-        $id = $datos->getIdEjemplar($slug);
-        $padres = $datos->getParents($id);
+            // return $padres;
+            foreach (DB::table('razas')->select('raza')->cursor() as $allRazas) {
+                $razas[$allRazas->raza] = $allRazas->raza;
+            }
 
-        // return $padres;
-        foreach (DB::table('razas')->select('raza')->cursor() as $allRazas) {
-            $razas[$allRazas->raza] = $allRazas->raza;
+            $columns = field::all();
+
+            $details = $datos->getDetails($slug);
+
+            return view('admin.ejemplars.editar-ejemplar', compact('details', 'razas', 'columns', 'padres'));
+        } else {
+            return redirect('/g/iniciar-sesion');
         }
-
-        $columns = field::all();
-
-        $details = $datos->getDetails($slug);
-
-        return view('admin.ejemplars.editar-ejemplar', compact('details', 'razas', 'columns', 'padres'));
     }
 
     /**
@@ -231,43 +286,48 @@ class EjemplarController extends Controller
 
     public function update(Request $request, $id)
     {
-        $datos = [];
-        $relation = new Relation();
-        $ejemplar = new Ejemplar;
-        $media = new Media;
+        $user = Auth::user();
+        if ($user) {
+            $datos = [];
+            $relation = new Relation();
+            $ejemplar = new Ejemplar;
+            $media = new Media;
 
-        $id = $ejemplar->getIdEjemplar($id);
-        $media->saveMedia($request, $id);
+            $id = $ejemplar->getIdEjemplar($id);
+            $media->saveMedia($request, $id);
 
-        // Ciclo for para actualizar los datos
-        foreach ($request->request as $key => $value) {
+            // Ciclo for para actualizar los datos
+            foreach ($request->request as $key => $value) {
 
-            // Actualizar relaciones condicionada para evitar errores
-            if ($key == "id_macho" || $key == "id_hembra") {
-                if (!is_null($value)) {
-                    $type = $key == "id_macho" ? 1 : 2;
-                    $idPadre = $ejemplar->getIdEjemplar($value);
-                    $relation->updateRelations($type, $idPadre, $id);
+                // Actualizar relaciones condicionada para evitar errores
+                if ($key == "id_macho" || $key == "id_hembra") {
+                    if (!is_null($value)) {
+                        $type = $key == "id_macho" ? 1 : 2;
+                        $idPadre = $ejemplar->getIdEjemplar($value);
+                        $relation->updateRelations($type, $idPadre, $id);
+                    }
+                }
+
+                $idColumn = DB::table('fields')
+                    ->where('name', $key)
+                    ->value('id');
+
+                /**
+                 * Actualizar datos condicionado para evitar errores
+                 * se actualiza de la siguiente manera field_id -> nuevoValor -> ID_Ejemplar
+                 */
+                if (!is_null($idColumn)) {
+                    data_field::where('field_id', $idColumn)
+                        ->where('ejemplar_id', $id)
+                        ->update(['data' => $value]);
+
                 }
             }
 
-            $idColumn = DB::table('fields')
-                ->where('name', $key)
-                ->value('id');
-
-            /**
-             * Actualizar datos condicionado para evitar errores
-             * se actualiza de la siguiente manera field_id -> nuevoValor -> ID_Ejemplar
-             */
-            if (!is_null($idColumn)) {
-                data_field::where('field_id', $idColumn)
-                    ->where('ejemplar_id', $id)
-                    ->update(['data' => $value]);
-
-            }
+            return redirect('/Ejemplar')->with('status', 'Ejemplar actualizado!');
+        } else {
+            return redirect('/g/iniciar-sesion');
         }
-
-        return redirect('/Ejemplar')->with('status', 'Ejemplar actualizado!');
     }
 
     /**
@@ -276,26 +336,41 @@ class EjemplarController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Request $request, $id)
+    public function destroy(Request $request, $slug)
     {
-        $accion = new Ejemplar();
-        $id = $accion->getIdEjemplar($id);
+        $user = Auth::user();
+        if ($user) {
+            $accion = new Ejemplar();
+            $id = $accion->getIdEjemplar($slug);
 
-        $ejemplar = Ejemplar::find($id);
-        $ejemplar->delete();
+           
 
-        relation::where('padre_id', $id)
-            ->update(['padre_id' => 0]);
+            relation::where('padre_id', $id)
+                ->update(['padre_id' => 0]);
 
-        $accionGet = new data_field();
-        $ejemplares = $accionGet->getDataField();
+            $accionGet = new data_field();
+            $media = new Media();
+            $delete = new MediaController();
+            $count = $media->getMedia($slug);
 
-        $ejemplares = $accionGet->paginador($ejemplares, null, 10);
+            if (count($count) > 1) {
+                foreach ($count as $key) {
+                    $delete->destroy($key->src, true);
+                }
+            }
+            
+            $ejemplar = Ejemplar::find($id);
+            $ejemplar->delete();
+
+            $ejemplares = $accionGet->getDataField();
+            $ejemplares = $accionGet->paginador($ejemplares, null, 10);
             $admin = true;
 
-            
-
-        return response()->json(view('public.sublista', compact('ejemplares', 'admin'))->render());
+          
+            return response()->json(view('public.sublista', compact('ejemplares', 'admin'))->render());
+        } else {
+            return redirect('/g/iniciar-sesion');
+        }
 
     }
